requirements.txt

langchain                
langchain-community      
langchain-core           
langchain-openai         
langchain-text-splitters 
langchain-pinecone                          
numpy                   
openai                              
pandas                                  
PyPDF2                          
python-dotenv         
streamlit                  